Added back-end logic for OwnerClass and EmployeeClass, added Feedback logic and Transactions logic.

+++ Core/Models/Order.cs
@@ -6,7 +6,7 @@ namespace Models
     public class Order
     {
         public int Id { get; set; }
-        public float TotalSum { get; set; }
+        public decimal TotalSum { get; set; }
         public List<MenuPosition> MenuPositions {get;set;}
         public int? UserId { get; set; }
         public DateTime DeliveryTime { get; set; }
